backend/requirements.txt
# Core dependencies
google-cloud-storage>=2.10.0
google-cloud-texttospeech>=2.16.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-api-python-client>=2.100.0

# Video processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Web framework for UI
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
